pyproject.toml
[project]
name = "ungraph"
version = "0.1.4"
description = "Python framework for building Knowledge Graphs from unstructured text using Neo4j (Cypher), aligned with property-graph standards"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Alejandro Giraldo Londoño", email = "alejandro@qnow.tech"}
]
license = {text = "MIT"}
keywords = ["knowledge-graph", "rag", "langchain", "neo4j", "unstructured-data"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "chardet>=5.2.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-docling>=2.0.0",
    "langchain-experimental>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-neo4j>=0.6.0",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "markdown>=3.10",
    "neo4j>=5.28.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "unstructured>=0.18.21",
]

[project.optional-dependencies]
# Inference - Para fase de inferencia con spaCy NER
# Incluye spaCy base y modelos de idioma
infer = [
    "spacy>=3.8.11",
]

# Inference con modelo de inglés
infer-en = [
    "ungraph[infer]",
    # Nota: Los modelos de spaCy se instalan con: python -m spacy download en_core_web_sm
    # Esto se documenta en la instalación, no se puede especificar en pyproject.toml
]

# Inference con modelo de español
infer-es = [
    "ungraph[infer]",
    # Nota: Los modelos de spaCy se instalan con: python -m spacy download es_core_news_sm
    # Esto se documenta en la instalación, no se puede especificar en pyproject.toml
]

# Inference con ambos idiomas (inglés y español)
infer-all = [
    "ungraph[infer-en,infer-es]",
]

# Graph Data Science - Para funcionalidades avanzadas de análisis de grafos
gds = [
    "graphdatascience>=1.18",
]

# Visualización de grafos en Jupyter
ynet = [
    "yfiles-jupyter-graphs-for-neo4j>=1.7.0",
]

# Herramientas de desarrollo
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.10",
    "matplotlib>=3.10.8",
    "pip>=25.3",
]

# Experimentos y evaluación
experiments = [
    "opik>=1.9.66",
]

# Todas las extensiones
all = [
    "ungraph[gds,ynet,dev,experiments,infer-all]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Configurar para empaquetar ungraph/ como el paquete ungraph
[tool.hatch.build.targets.wheel]
packages = ["ungraph"]
# Excluir directorios que no deben estar en el paquete instalable
exclude = [
    "ungraph/notebooks",
    "ungraph/pipelines",
    "ungraph/data",
    "ungraph/scripts",
    "ungraph/main.py"
]

# Configuración para TestPyPI
[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
default = false

# Índice por defecto (PyPI oficial)
[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
publish-url = "https://upload.pypi.org/legacy/"
explicit = false
default = true


